kde-plasma/libkworkspace: New package (split from plasma-workspace)
In Plasma 5.13, kde-cli-tools began to depend on libkworkspace for
Wayland support in KCMs. Besides making circular dep workarounds
necessary, this also meant a considerable chunk of Plasma added to
the kde-cli-tools depgraph, which is used by applications even if
not having a Plasma desktop installed.

Revdeps already search for libkworkspace independently, which means
no revdep patching is required to support this split package.

Signed-off-by: Andreas Sturmlechner <asturm@gentoo.org>
Package-Manager: Portage-2.3.51, Repoman-2.3.11